# Faster multi-file matching