--- a/go.mod
+++ b/go.mod
@@ -3,7 +3,6 @@ module github.com/hmarr/codeowners
 go 1.14
 
 require (
-	github.com/karrick/godirwalk v1.17.0
 	github.com/spf13/pflag v1.0.5
 	github.com/stretchr/testify v1.8.0
 )